Add settings functionality
