+++ server/getVaccinationType.js
@@ -1,0 +1,61 @@
+module.exports = function getVaccinationType(string) {
+  const cases = {
+    rotavirusG1: '1. Grundimmunisierung',
+    rotavirusG2: '2. Grundimmunisierung',
+    rotavirusG3: '3. Grundimmunisierung',
+    rotavirusG4: '4. Grundimmunisierung',
+    tetanusG1: '1. Grundimmunisierung',
+    tetanusG2: '2. Grundimmunisierung',
+    tetanusG3: '3. Grundimmunisierung',
+    tetanusG4: '4. Grundimmunisierung',
+    tetanusA1: '1. Auffrischimpfung',
+    tetanusA2: '2. Auffrischimpfung',
+    tetanusA: 'Auffrischimpfung',
+    diphtherieG1: '1. Grundimmunisierung',
+    diphtherieG2: '2. Grundimmunisierung',
+    diphtherieG3: '3. Grundimmunisierung',
+    diphtherieG4: '4. Grundimmunisierung',
+    diphtherieA1: '1. Auffrischimpfung',
+    diphtherieA2: '2. Auffrischimpfung',
+    diphtherieA: 'Auffrischimpfung',
+    pertussisG1: '1. Grundimmunisierung',
+    pertussisG2: '2. Grundimmunisierung',
+    pertussisG3: '3. Grundimmunisierung',
+    pertussisG4: '4. Grundimmunisierung',
+    pertussisA1: '1. Auffrischimpfung',
+    pertussisA2: '2. Auffrischimpfung',
+    pertussisA3: '3. Auffrischimpfung',
+    hibG1: '1. Grundimmunisierung',
+    hibG2: '2. Grundimmunisierung',
+    hibG3: '3. Grundimmunisierung',
+    hibG4: '4. Grundimmunisierung',
+    poliomyeitisG1: '1. Grundimmunisierung',
+    poliomyeitisG2: '2. Grundimmunisierung',
+    poliomyeitisG3: '3. Grundimmunisierung',
+    poliomyeitisG4: '4. Grundimmunisierung',
+    poliomyeitisA1: '1. Auffrischimpfung',
+    hepatitisBG1: '1. Grundimmunisierung',
+    hepatitisBG2: '2. Grundimmunisierung',
+    hepatitisBG3: '3. Grundimmunisierung',
+    hepatitisBG4: '4. Grundimmunisierung',
+    pneumokokkenG1: '1. Grundimmunisierung',
+    pneumokokkenG2: '2. Grundimmunisierung',
+    pneumokokkenG3: '3. Grundimmunisierung',
+    pneumokokkenG4: '4. Grundimmunisierung',
+    pneumokokkenS: 'Standardimpfung',
+    meningokokkenG1: '1. Grundimmunisierung',
+    masernG1: '1. Grundimmunisierung',
+    masernG2: '2. Grundimmunisierung',
+    masernS: 'Standardimpfung',
+    mumpsG1: '1. Grundimmunisierung',
+    mumpsG2: '2. Grundimmunisierung',
+    variazellenG1: '1. Grundimmunisierung',
+    variazellenG2: '2. Grundimmunisierung',
+    hpvG1: '1. Grundimmunisierung',
+    hpvG2: '2. Grundimmunisierung',
+    herpesZosterG1: '1. Grundimmunisierung',
+    herpesZosterG2: '2. Grundimmunisierung',
+    influenza: 'Standardimpfung',
+  }
+  return cases[string]
+}
